Re-implement moving of permissions to a core menu item
